Move previews to androidMain
