+++ composeApp/build.gradle.kts
@@ -10,6 +10,7 @@ kotlin {
         androidMain.dependencies {
             implementation(libs.koin.android)
             implementation(libs.lottie)
+            implementation(compose.preview)
         }
         commonMain.dependencies {
             implementation(projects.featureListStreams)
@@ -26,7 +27,6 @@ kotlin {
 
             implementation(compose.material3)
             implementation(compose.ui)
-            implementation(compose.preview)
 
             implementation(libs.koin.core)
             implementation(libs.koin.compose)
