+++ core-shared-ui/src/androidMain/kotlin/com/codandotv/streamplayerapp/core_shared_ui/widget/StreamPlayerTopBarPreview.kt
@@ -1,0 +1,24 @@
+package com.codandotv.streamplayerapp.core_shared_ui.widget
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberTopAppBarState
+import androidx.compose.runtime.Composable
+import com.codandotv.streamplayerapp.core_shared.extension.empty
+import com.codandotv.streamplayerapp.core_shared_ui.theme.ThemePreview
+import com.codandotv.streamplayerapp.core_shared_ui.theme.ThemePreviews
+
+@OptIn(ExperimentalMaterial3Api::class)
+@ThemePreviews
+@Composable
+fun StreamPlayerTopBarPreview() {
+    ThemePreview {
+        StreamPlayerTopBar(
+            scrollBehavior = TopAppBarDefaults.exitUntilCollapsedScrollBehavior(
+                rememberTopAppBarState()
+            ),
+            onNavigateProfilePicker = {},
+            onSelectedProfilePicture = String.empty()
+        )
+    }
+}
