+++ core-shared-ui/src/commonMain/kotlin/com/codandotv/streamplayerapp/core_shared_ui/widget/StreamsCard.kt
@@ -10,7 +10,6 @@ import androidx.compose.material3.Card
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.codandotv.streamplayerapp.core_shared.Url.IMAGE_URL_SIZE_300
@@ -50,15 +49,3 @@ data class StreamsCardContent(
     val url: String,
     val contentDescription: String,
 ) : Parcelable
-
-@Preview
-@Composable
-fun StreamsCardPreview() {
-    StreamsCard(
-        StreamsCardContent(
-            url = "${IMAGE_URL_SIZE_300}evgwd37VHBJhXvSr88Mrx5riFil.jpg",
-            contentDescription = "Test 1",
-            id = "",
-        )
-    )
-}
