+++ core-shared-ui/src/commonMain/kotlin/com/codandotv/streamplayerapp/core_shared_ui/widget/StreamsCarousel.kt
@@ -12,7 +12,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.paging.PagingData
@@ -69,14 +68,3 @@ data class StreamsCarouselContent(
     val genreTitle: String,
     val contentList: Flow<PagingData<StreamsCardContent>>
 )
-
-@Composable
-@Preview
-fun StreamsCarouselPreview() {
-    StreamsCarousel(
-        content = StreamsCarouselContent(
-            genreTitle = "Ação",
-            contentList = emptyFlow()
-        )
-    )
-}
