+++ feature-search/src/androidMain/kotlin/com/codandotv/streamplayerapp/feature_search/presentation/widgets/StreamsCarouselPreview.kt
@@ -1,0 +1,19 @@
+package com.codandotv.streamplayerapp.feature_search.presentation.widgets
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+@Preview
+fun StreamsErrorPreview() {
+    StreamsError(
+        onRetry = {},
+        onCloseButton = {}
+    )
+}
+
+@Composable
+@Preview
+fun StreamEmptyPreview() {
+    StreamsEmpty()
+}
